==> processedData/spatial/vms/intermediate/01_process_and_clean_VMS_2009_2016.R <==
# This script will clean the Processed VMS csv files 2009-2016 received in Sep 2016
# also see blake's xlsx file "VMS csv processing summary"

# Remove all garbage headers from each .txt file
# Ensure each column is formatted properly (date-time, etc)
# Spit out clean RDS and csv files to be used instead of raw .txt files

library(stringr)

# load files
setwd("/Users/jameal.samhouri/Documents/CNH_to_github/cnh/")
setwd("/Users/jamealsamhouri/Documents/cnh/")

# write a function to open each raw VMS data file and do bare bones cleaning.
# expect wanring messages about coercing NAs

#filename <- "VMSNWFS16-007_2010" #

bare_bones <- function(filename){
#   
#   # Remove all garbage headers from each .txt file
#   
#   # read the number of dashes to determine how to break up columns
# #  dtemp <- read.table(paste0("rawData/VMS/Processed VMS csv files 2009-2016/",filename,".csv"), skip=1,nrow=1)
#   #my_widths <- as.vector(apply(dtemp,1,nchar)+1)
#   
#   #my_widths <- c(21,11,11,29,11,11,21,16)
#   
#   # determine how many rows to skip
#   x = 0
#   bad_dash = TRUE
#   while(bad_dash){
#     dtemp <- read.csv(paste0('rawData/VMS/Processed VMS csv files 2009-2016/', filename, '.csv'), skip = x, nrow = 1,
#                         stringsAsFactors = FALSE)
#     old_length = nchar(dtemp[1])
#     new_length = nchar(gsub('-',"",dtemp[1]))
#     bad_dash = ifelse(old_length!=new_length, FALSE, TRUE)
#     x = x + 1
#   }
#   
#   my_widths <- as.vector(apply(dtemp, 1, nchar)+1)
#   
#   dat1 <- read.fwf(paste0("rawData/VMS/Processed VMS csv files 2009-2016/",filename,".csv"),widths=my_widths,skip=x)
#   #head(dat1,40)
 
#  dat1a <- apply(dat1, 2, function (x) str_trim(x, side="both"))
  
  dat1a <- read.csv(paste0('rawData/VMS/Processed VMS csv files 2009-2016/', filename, '.csv'), stringsAsFactors = FALSE)
  #head(dat1a)
  
  colnames(dat1a) <- c("DateTime", "Latitude", "Longitude", "VesselName", "AvgSpeed", "AvgCourse", "DocNum", "Declarations")
  dat1b <- data.frame(dat1a, stringsAsFactors = FALSE)
  #head(dat1b,40)
  
  # Ensure each column is formatted properly (date-time, etc)
  
  # format DateTime column
  dat1b$DateTime <- as.POSIXct(dat1b$DateTime, format = "%Y.%m.%d %H:%M", tz="Etc/GMT-8")
  #head(dat1b,40)
  
  # convert Lat-Long columnbs to numeric
  dat1b$Latitude <- as.numeric(dat1b$Latitude)
  dat1b$Longitude <- as.numeric(dat1b$Longitude)
  # head(which(is.na(dat1b$Latitude)))
  # dat1b[3042857:3042859,]
  # dat1b[235:237,]
  # dat1b[271:273,]
  
  # remove rows with NAs for Lat-Long (we think they occur when vessel names spilloveror a few boats have error reporting)
  
  lat_problems <- length(which(is.na(dat1b$Latitude))) # added on 9/16/16 to deal with situations where none of the latitudes are NA
  if(lat_problems>0) dat1b <- dat1b[-which(is.na(dat1b$Latitude)),]
  
  # head(dat1b,40)
  
  # check that NAs for lat-long have been removed
  head(which(is.na(dat1b$Longitude)))
  
  # make AvgSpeed and AvgCourse columns numeric
  dat1b$AvgSpeed <- as.numeric(dat1b$AvgSpeed)
  dat1b$AvgCourse <- as.numeric(dat1b$AvgCourse)
  
  # change N/As to NAs for Declarations column
  head(which(is.na(dat1b$Declarations)))
  dat1b$Declarations <- as.numeric(dat1b$Declarations)
  
  saveRDS(dat1b, paste0("rawData/VMS/Processed VMS csv files 2009-2016_clean/",filename,"_clean.RDS"))
  write.csv(dat1b,paste0("rawData/VMS/Processed VMS csv files 2009-2016_clean/",filename,"_clean.csv"), row.names=FALSE)
  
}

#system.time({bare_bones(filename)})

# read in filenames in directory
fn <-dir("rawData/VMS/Processed VMS csv files 2009-2016")
fn2 <- gsub(".csv","",fn)

# create clean RDS and .csv files for all raw .txt files

for(i in 1:length(fn2)){
  
  bare_bones(fn2[i])
  
}
